@@ -4,13 +4,12 @@
   <artifactId>studsluzba_server</artifactId>
   <version>0.0.1-SNAPSHOT</version>
   <properties>
-    <maven.compiler.source>11</maven.compiler.source>
-    <maven.compiler.target>11</maven.compiler.target>
+	  <java.version>21</java.version>
   </properties>
   <parent>
         <groupId>org.springframework.boot</groupId>
         <artifactId>spring-boot-starter-parent</artifactId>
-        <version>2.2.2.RELEASE</version>
+        <version>3.4.4</version>
         <relativePath/> <!-- lookup parent from repository -->
   </parent>
   <dependencies>	
@@ -25,6 +24,7 @@
 	<dependency>
 		<groupId>mysql</groupId>
 		<artifactId>mysql-connector-java</artifactId>
+		<version>8.0.33</version>
 		<scope>runtime</scope>
 	</dependency>
 	<dependency>
@@ -67,7 +67,11 @@
 	  <groupId>org.springframework.boot</groupId>
 	  <artifactId>spring-boot-starter-security</artifactId>
     </dependency>
-
+  	<dependency>
+	  <groupId>jakarta.validation</groupId>
+	  <artifactId>jakarta.validation-api</artifactId>
+	  <version>3.0.2</version>
+	</dependency>
   </dependencies>
 	<build>
         <plugins>
