@@ -2,7 +2,7 @@ package org.raflab.studsluzba.controllers.request;
 
 import lombok.Data;
 
-import javax.validation.constraints.NotEmpty;
+import jakarta.validation.constraints.NotEmpty;
 
 @Data
 public class DrziPredmetNewRequest {
