@@ -1,7 +1,7 @@
 package org.raflab.studsluzba.controllers.request;
 
 import lombok.Data;
-import javax.validation.constraints.NotNull;
+import jakarta.validation.constraints.NotNull;
 
 @Data
 public class ObnovaGodineInitRequest {
