@@ -6,7 +6,7 @@ import lombok.Data;
 
 import java.time.LocalDate;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 
 @Entity
 @Data
