@@ -10,7 +10,7 @@ import org.springframework.data.repository.CrudRepository;
 public interface PrijavaIspitaRepository extends CrudRepository<PrijavaIspita, Long> {
 	
 	
-	@Query("select pi from PrijavaIspita pi where pi.studentIndeks.id = :indeksId and pi.ispit.ispitniRok.skolskaGodina.aktivna = 1")
+	@Query("select pi from PrijavaIspita pi where pi.studentIndeks.id = :indeksId and pi.ispit.ispitniRok.skolskaGodina.aktivna = true")
 	List<PrijavaIspita> getPrijaveIspitaForIndeksAktivnaGodina(Long indeksId);
 	
 	@Query("select pi from PrijavaIspita pi where pi.studentIndeks.id = :indeksId and pi.ispit.id = :ispitId")
@@ -22,7 +22,7 @@ public interface PrijavaIspitaRepository extends CrudRepository<PrijavaIspita, L
 	@Query("select pi from PrijavaIspita pi where pi.ispit.id = :idIspita")
 	List<PrijavaIspita> getPrijaveZaIspit(Long idIspita);
 	
-	@Query("select pi from PrijavaIspita pi where pi.ispit.drziPredmet.skolskaGodina.aktivna=1 and pi.ispit.drziPredmet.nastavnik.id = :idNastavnik "
+	@Query("select pi from PrijavaIspita pi where pi.ispit.drziPredmet.skolskaGodina.aktivna=true and pi.ispit.drziPredmet.nastavnik.id = :idNastavnik "
 			+ "and pi.ispit.drziPredmet.predmet.id = :idPredmet and pi.ispit.ispitniRok.id = :idIspitniRok")
 	List<PrijavaIspita> getPrijavljeniForIspitNastavnikPredmetIspitniRokAktivna(Long idNastavnik, Long idPredmet, Long idIspitniRok); 
 	
