@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 public interface SkolskaGodinaRepository extends CrudRepository<SkolskaGodina, Long> {
 	
-	@Query("select s from SkolskaGodina s where s.aktivna=1")
+	@Query("select s from SkolskaGodina s where s.aktivna=true")
 	SkolskaGodina getAktivnaSkolskaGodina();
 
 	SkolskaGodina findTopByOrderByKrajnjaDesc();
